server-side/requirements.txt
fastapi
pydantic
uvicorn
alembic
python-dotenv
psycopg2